Move defaultBackStackRecordLocalProviders and ViewModelBackStackRecordLocalProvider to foundation
