+++ backstack/src/androidMain/kotlin/com/slack/circuit/backstack/BackStackRecordLocalProvider.android.kt
@@ -1,5 +1,1 @@
-package com.slack.circuit.backstack
 
-internal actual val defaultBackStackRecordLocalProviders:
-  List<BackStackRecordLocalProvider<BackStack.Record>> =
-  listOf(ViewModelBackStackRecordLocalProvider)
