+++ backstack/src/browserCommonMain/kotlin/com/slack/circuit/backstack/BackStackRecordLocalProvider.js.kt
@@ -1,5 +1,1 @@
-package com.slack.circuit.backstack
 
-internal actual val defaultBackStackRecordLocalProviders:
-  List<BackStackRecordLocalProvider<BackStack.Record>> =
-  emptyList()
