+++ backstack/src/jvmMain/kotlin/com/slack/circuit/backstack/BackStackRecordLocalProvider.jvm.kt
@@ -1,5 +1,1 @@
-package com.slack.circuit.backstack
 
-internal actual val defaultBackStackRecordLocalProviders:
-  List<BackStackRecordLocalProvider<BackStack.Record>> =
-  emptyList()
