+++ circuit-foundation/src/commonJvmTest/kotlin/com/slack/circuit/foundation/ProvidedValuesLifetimeTest.kt
@@ -61,7 +61,7 @@ class ProvidedValuesLifetimeTest {
             providedValues =
               providedValuesForBackStack(
                 backStack = backStack,
-                stackLocalProviders = persistentListOf(TestBackStackRecordLocalProvider),
+                backStackLocalProviders = persistentListOf(TestBackStackRecordLocalProvider),
               ),
           )
         }
