+++ circuit-foundation/src/jvmMain/kotlin/com/slack/circuit/foundation/internal/DefaultBackStackRecordLocalProviders.jvm.kt
@@ -1,0 +1,10 @@
+// Copyright (C) 2025 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuit.foundation.internal
+
+import com.slack.circuit.backstack.BackStack
+import com.slack.circuit.backstack.BackStackRecordLocalProvider
+
+internal actual val defaultBackStackRecordLocalProviders:
+  List<BackStackRecordLocalProvider<BackStack.Record>> =
+  emptyList()
